+++ server/.eslintrc.json
@@ -1,3 +1,16 @@
 {
-  "extends": ["../.eslintrc.json"]
+  "env": {
+    "es2021": true,
+    "node": true
+  },
+  "plugins": ["jsdoc"],
+  "extends": "google",
+  "overrides": [],
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "rules": {
+    "valid-jsdoc": "warn"
+  }
 }
