+++ server/package.json
@@ -20,8 +20,12 @@
   },
   "devDependencies": {
     "@types/jest": "^29.4.0",
-    "eslint": "^8.35.0",
+    "eslint": ">=5.16.0",
+    "eslint-config-google": "^0.14.0",
+    "eslint-plugin-jsdoc": "^40.0.2",
     "jest": "^29.5.0",
-    "nodemon": "^2.0.21"
+    "jsdoc": "^4.0.2",
+    "nodemon": "^2.0.21",
+    "prettier": "^2.8.4"
   }
 }
